form edit

+++ app/Http/Controllers/ProjectController.php
@@ -31,7 +31,7 @@ class ProjectController extends Controller
         $pr->project_info = $request['project_info'];
 
         return ($pr->save() !== 1) ?
-            redirect('/projects')->with('status_success', 'Project created!') :
+            redirect('/projects')->with('status_success', 'Project "' . $request['project_name'] . '" created!') :
             redirect('/projects')->with('status_error', 'Project was not created!');
     }
 
@@ -46,13 +46,19 @@ class ProjectController extends Controller
         $pr->project_name = $request['project_name'];
         $pr->project_info = $request['project_info'];
         return ($pr->save() !== 1) ?
-            redirect('/projects/' . $id)->with('status_success', 'Project updated!') :
-            redirect('/projects/' . $id)->with('status_error', 'Project was not updated!');
+            redirect('/projects/' . $id)->with('status_success', 'Project "' . $request['project_name'] . '" updated!') :
+            redirect('/projects/' . $id)->with('status_error', 'Project "' . $request['project_name'] . '" was not updated!');
     }
 
     public function destroy($id)
     {
-        Project::destroy($id);
-        return redirect('/projects')->with('status_success', 'Project deleted!');
+        $project = Project::find($id);
+        $employee_count = $project->employees()->count();
+        if (!$employee_count > 0) {
+            Project::destroy($id);
+            return redirect('/projects')->with('status_success', 'Project "' . $project['project_name'] . '" deleted!');
+        } else {
+            return redirect('/projects')->with('status_error', 'Project "' . $project['project_name'] . '" can no be deleted. One or more employees are assigned to the project');
+        }
     }
 }
